fix(storage): index_id is an int64 now
Several others minor fixes:
* fix(cmake.sh): several bad lines fixed.
* fix(sonarqube): no need to analyze ut.xml
* fix(storage): INFINITY is only float, replaced by std::numeric_limits...

REFS: MON-10966

+++ storage/inc/com/centreon/broker/storage/conflict_manager.hh
@@ -343,7 +343,7 @@ class conflict_manager {
 
   int32_t send_event(stream_type c, std::shared_ptr<io::data> const& e);
   int32_t get_acks(stream_type c);
-  void update_metric_info_cache(uint32_t index_id,
+  void update_metric_info_cache(uint64_t index_id,
                                 uint32_t metric_id,
                                 std::string const& metric_name,
                                 short metric_type);
